requirements.txt
# Core Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# Tracking
filterpy>=1.4.5
scikit-learn>=1.3.0
lap>=0.4.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0